requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bokeh==3.3.4
contourpy==1.2.0
Django==4.2.8
django-bootstrap-v5==1.0.11
django-rest-framework==0.1.0
djangorestframework==3.14.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
tornado==6.4
tzdata==2023.4
xyzservices==2023.10.1
